Import long from past instead of int from future
This reverts commit 77dcc1107c2f3351589da34b3be3a6881b62c921 "Use int from
futurize instead of long" and replaces the solution by leaving long as is and
importing it from past.builtins.  This was necessary as it revealed, that in
Python 2 in one file we do:

from builtins import int
x = int(y)

and pass x to a method deefined in another file where int is not imported:

assert isinstance(x, int)

This raises AssertionError because these two int types are not compatible.
Even if we added the int import into every file, we risk getting values from
outside.  Thus we need to leave long as is in Python 2 and import it from past
for Python 3 compatibility.