Make remove_duplicates work on Python 3
Python 3 doesn't support sorting heterogenopus lists.  Thus the
remove_duplicates algorithm using sorting fails on Python 3 when the list
contains items of different types.  Also the argument keep_order is never used,
so we can remove it.  Using a set to remove the duplicates seems more than
twice more performant than the original algorighm.  This makes the
implementation so trivial, that it would be better to inline this code and
remove the function 'remove_duplicates()' from utils.  But because it is used
in applications, we can't afford it right now.