Make __str__ methods Python 3 copmpatible
Use the @python_2_unicode_compatible decorator to make all __str__ methods
Python 3 compatible.  What it does is that in Python 2 this renames __str__ to
__unicode__ and implements __str__ as unicode(self).encpode('utf-8').  Thus we
have both methods covered in Python 2 from one definition and in Python 3 there
is no __unicode__ any more so nothing needs to be done.  So where we used to
have __unicode__, it is now renamed to __str__.

The commit also attempts to move the __str__ method (and oftent also other
__*__ methods) to the top of the class definition because this is a good
convention.