Prevent DB error during row logging
When running tests, logging the fetched row in DBDataPostgreSQL.fetch() caused
DBSystemException because of uninitialized DB connection within the
stringification of one of the row's column's enumerator.  Hard to say why the
enumerator used invalid connection parameters, but it is clear that by logging
a row we surely don't want to initialize all its enumerators.