Fix multiform side form referencing
The previous implementation using redundant lists of forms and their tab order
got out of sync time to time and it lead to strane problems, such as invoking
commands on other than the current side form.  The user for example wished to
insert a record into the side form, but got an insertion form for the side form
next to the current one.

After unsuccesfull attempts to solve various corner cases of the prevoius
solution I am ditching the whole mystery of redundant lists and implementing
all operations purely through the AuiNotebook methods, which are totally
sufficient for what we need.  It seems much cleaner and I don't see any
problems any more.  It is, however a larger change so testing recommended!