Show newly added side forms automatically
Side forms newly added to the specification were not displayed automatically to
users having customized their tab order or hidden some tabs.  This was because
it was not possible to determine which side forms were previously hidden and
which are completely new.  Now we save the hidden forms together with the
customized tab order so we are able to recognize the new forms on the next run
and display them to the user.