Don't discard SystemExit error messages
We need to re-raise the exception in order to see the error message.  The
previous solution worked well for situations when SystemExit does not include
an error message, but this is not always the case.  For example the error
"Unable to access the X Display, is $DISPLAY set properly?" produced by
wxPython initialization got discarded resulting in major confusion.