Avoid gsql introspection on app startup
inspect.stack() causes a serious delay on application startup.  It is less
serious on Python 2 but one call may take even a second on Python 3.  The
filename and line number is only important when running gsql and is useless in
application runtime, thus don't save it unless we are running gsql.py.