Lock DBConfig on update
This commit is an attempt to fix #4.  The assumption is that there must be some
conflicting selectin in progress when update() causes
self._pg_select_transaction being closed in _pg_begin_transaction().  And
because DBConfig already locks on selection, we can use the same lock for
update and possibly avoid the conflict.