Don't use deprecated "oid" columns in tables and views
The 'oid' columns are deprecated for a long time (see also
pytis.data.deprecated).
Until now, they were supported by postgresql, but it is allready
changed in Postgresql12. So it is necessary to update
all table and view definitions containing "oid" columns,
otherwise it would cause problems when upgrading to Postgresql12.