New infrastructure for Pytis API access from applications
Current applications access Pytis APIs directly.  This results in a mess as the
Pytis public API is designed mainly to be accessed from within Pytis itself.
Having two very different consumers of one API is a problem.  Thus we attempt
to separate the two APIs completely.  The new API accessed through
pytis.api.app is specifically designed for the applications.  All pytis APIs
not explicitly marked by pytis.api decorators are now considered to be
internal.

We will need to modify the applications to use this new API.  This will often
require extension of this API.  The final goal is to completely stop using the
internal Pytis APIs from all applications.