Move pytis.form.application._appplication to pytis.form.app
We need to access the application's methods from all around Pytis and having
the app instance private lead to duplication of many of its methods as
functions within the pytis.form.application module.  These functions are
impractical, unnecessary and cauase a lot of clutter.  The goal is to gradually
remove all these functions and access them as methods of the app object.

Note the difference between pytis.api.app (API for applications) and
pytis.form.app (internal API for Pytis).  As pytis.forms should not be used by
apps at all (in future when we implement all necessary features through the
API), this should hopefully not cause much confusion.