Separate internal/public namespace in API definition
Adding new public API methods started to cause troubles because of conflicts
with existing internal API method/attribute names.  This commit adds the api_
prefix to public API definitions (which is stripped in the final API) and thus
allows to have the same name for internal use and for the public API with
different meaning (where necessary).  For example we prefer properties in the
public API and we already have a method of the same name in the internal API.
This way we don't need to change the internal API in order to add the property
to the public API.