Ask all export questions at once
From the user's perspective it is better to ask all questions consistently at
once than asking a series of question and changing the series according to
conditions.  It feels too magical and the user is not sure how it works.

From the code perspective, we don't want to duplicate the same code twice.