Remove remote functions library_version() and x2goclient_version()
These functions are not used anywhere in the applications and pytis (except for
the now removed check which seems useless).  There are too many version
information sources which makes it all very confusing.  We are now keeping just
one version source, which is initialized in _connect() and saved in RPCInfo
instance.  It is obtained through 'x2goclient_version', but does not need a
saparate function in pytis.remote as it is kept in RPCInfo instance.