Work around broken modality after closing a codebook form
Modality was broken after closing a modal dialog (such as a codebook form)
which was opened above another modal dialog (such as an edit form).  Before
this change, it was possible to select rows in the top level browse form even
though the modal edit form was stil open.

Relates to issue #4.