New argument filetypes of pytis.remote.open_selected_file() & comp.
We plan to replace the arguments 'patterns' and 'pattern' by 'filetypes'.  This
commit is the first transitional stage, where we allow adjusting the
applications to use 'filetypes' instead of patterns, while patterns stil work
as before.  Once the applications are updated, we'll be able to remove the old
arguments from the public API.