Don't change selection when field focus lost
The field selection is now always removed and the cursor is moved to the first
character whenever the field loses focus.  This happens for example during
switching between windows (such as between a pytis form and an invoice from
which the user wants to enter some data into the form).  The selection/cursor
change in this case is very unpleasant as the user doesn't expect it and needs
to manually move the cursor back to the end each time.

The selection reset hack was there originally to aviod the problem of invisible
field content when the selection was gray/inactive (lost focus) and (as the
comment suggests) when the field background color was set.  We used to change
field background colour to indicate invalid fields, but we don't do that any
more (we have the validation icons).  The only background color changes are now
performed to indicate field editability.  The invisible content problem doesn't
seem to happen now neither for editable or ineditable fields.  Thus I don't see
any reason to keep the selection hack when it actually causes other problems.