Add basic support for PostgreSQL JSON data types
The PostgreSQL types JSON and JSONB can now be used in DB specifications and
the DB interface handles them as expected.  The new pytis data types JSON and
JSONB work with a Python dictionary or list as its internal value.  Primitive
values are not allowed as:
1) it is useless to represent them as JSON (we have specific types for them)
2) it would be problematic to distinguish between None as a real value and None
as empty value (NULL in the database).
3) it would need some refactoring to make it possible to correctly pass the
JSON values to psycopg2.
Thus any JSON value in Pytis must be a Python dict or list (tuple is accepted
too, but represented as JSON array (same as Python list).

Establishes base ground for solving issue #14.