Fix client connection detection logic
The previous logic was incorrect in case of running the app within an X2Go
session, which is not a Pytis2Go session.  Such session should be considered as
if we were running locally, because the RPyC communication can not be
established.  Instead of relying only on client_ip(), we also need to detect
whether the pytis info file (pytis_x2go_info_file()) exists (or existed at
least once since the app has been started).  To achieve that we need another
function in addition to 'client_ip()' and 'client_available()', which will
detect the presence of the pytis info file.  I decided to change the meaning of
'client_available()' to return True if the connection is *possibly*
available (this is an X2Go session and the pytis info file exists) and rename
the previous 'client_available()' function to 'client_connection_ok()' which, I
believe, is more intuitive...