Fix Pytis CMS dbdefs inclusion mechanism
The previous mechanism was based on overriding the Pytis CMS spcification
classes, but this caused a lot of troubles because of the way how gsql works.
One consequence of this was the need to redefine the names of the original DB
objects using a leading undersore to avoid duplicate names (refused by gsql).
Even redefining the names (which required to redefine them back in derived
classes) didn't end up the series of hacks necessary to make it work.  It was
also necessary to redefine other attributes such as 'depends_on' or
'insert_order' which refer to the non-overriden versions of the DB objects.  In
the end, it was not possible to make the definition of Pytis CMS in Pytis Demo
actually work.

The new approach is based on changing the original Pytis CMS DB object
definitions in place.  This way we don't get any duplicate objects and avoid
the need for further hacks dealing with them.

Beware: As this commit changes the supposed way of Pytis CMS embedding, any
projects which use Pytis CMS need to be updated!