Decouple browser resource server API logic
Avoid unnecessary APIs between the exporter, browser and resource server.  This
simplifies the whole logic and makes the code more readable.  The changes
actually come from the attempt to rewrite the resource server as a
'wx.html2.WebViewHandler' which proved to be problematic due to the lack of
connection between the handler and browser instances (there is just one global
handler instance while we have one resource server instance per browser
instance).