Use original table column doc and label for DB views
The label and doc were correctly used only when the specification was based on
a DB table.  DB view columns don't carry the information from the underlying
table automatically.  Thus we need to find out the original table column
specification from the information we have in '.original_column()' and use it
to get the defalut value for the label and descr of a specification 'Field'
instance.

Relates to issue #54.