Avoid calling pytis.form.top_level_exception() in postgresql.py
Calling pytis.form.top_level_exception() sems an unnecessary hack here.  The
motivation was most likely to force falling into the bug report dialog inside
the application (which is normally not the case for DBUserException).  But why
all this hassle when ValueError does exactly the same and seems to be a good
fit for this kind of error anyway?

Using DBUserException actually caused some errors in tests to be quite hard to
track because of special treatment of DBException at various places.