Completely rewrite support for command definition and handling
This rewrite addresses several issues:

1. Defining menu items in applications needs to be done through Pytis internals -
classes and commands defined in pytis.form.  To avoid this, we need the ability
to define commands through the public API.  As menu items mostly run forms or
code, it would make sense to use the existing API for these operations, such as
'app.run_form()' and 'app.run_procedure()'.  We also need a clear distinction
which commands are public and which are internal.

2. Commands are defined dynamically by class constants, such as
'Application.COMMAND_RUN_FORM'.  This practice reveals to be problematic in
combination with linters and code analysis tools as these constants are not
defined statically so these tools don't know about them.  There is also no
explicit connection between these commands and the methods, which handle them,
such as 'Application._cmd_run_form()'.

3. The commands currently lack usable documentation.  The documentation at
command definition is hard to manage as it is "distant" from the actual code,
which performs the command.

4. User actions are currently expressed as a pair of Command instance and its
arguments.  This pair is now passed all around the application as a tuple,
which proved to be very tedious and impractical for many reasons.  Having a
single instance encapsulating the command and its arguments would be more
practical.

The new implementation resolves these issues and adds some benefits.  The new
"infrastructure" is defined independently of the user interface classes and
thus is fully testable.  Most tests are written as doctests, which also makes
the key principals more obvious and discoverable.  Command handling methods are
explicitly marked and documented at the place of their definition.  Command
arguments can be easily introspected and their compatibility with command
method arguments can be verified.  Command can be directly bound to a
particular user interface element instance, which was before solved by an ugly
hack with adding the '_command_handler' argument.  The new solution is clearly
more Pythonic and elegant.  Command instance definition is now often more space
efficient to write, etc.