Display dialog when locked row editation fails
When a user attempts to open an edit form on a row which is currently locked,
the form is not opened.  The situation was previously handled by displaying a
message "The record is locked." in the status bar.  The problem was, that the
message was immediately deleted thanks to clearing the status bar after closing
a form (which is questionable by itself).  Anyway, the status bar message
doesn't seem enough for the situation.  An interactive dialog box, which needs
to be confirmed by the user, seems more appropriate.  It may also contain some
more information for the user.

Related to #71.