Stop attempts to get lock tables by parsing query tree
As discussed in #71, parsing the query tre currently doesn't seem to work at
all and it causes a significant delay in complicated views so let's completely
leave this path.  If we want to find out the tables to lock automatically, we
can inspect the sqlalchemy view definition (which was not avialable at the time
of writing the query tree parsing functionality).