Add support for mode and encoding in app.open_selected_file()
These arguments were not supported in Pytis API because ExposedFileWrapper
doesn't support them on old Pytis2go clients.  This commit emulates text
decoding at the level of Application._ExposedFileWrapper (the wrapper of
ExposedFileWrapper which we need anyway).  The behavior is transparent for the
application and we don't need to handle text decoding in application code any
more.