Improve batch insertion response when saved
Announcing saved record through app.echo() is not enough because the message is
typically immediately overriden by validation massages after loading the next
record so the user may not be sure whether he is stil at the previous record or
already at the next one.  Interactive messages thus seems to be justified.