Support newer versions of pymupdf
In order to use Pytis with Python3 we need to make some
adjustments when working with fitz module. But we also
need to keep the obsolete versions of pymupdf working
because Python2 does not support current versions of pymupdf.
So the adjustments are a bit ugly but we can remove them
as soon as we switch completely to Python3.