Make statusbar help work for popup menus
I got this long standing bug finally solved by a little hack.  After some
debugging I realized, that the help text actually is displayed in the
statusbar, but is immediately overwritten by some wx intenal call by an empty
text and thus can not be seen.  But it turned out that this problem olny
appears in the first statusbar field.  This field has a special meaning for wx.
So the hack was to add a dummy first field with zero width to the statusbar and
use the second field for the echo area.