Fix pytis.extensions menu check functions
The reasons to make this change:
1) get_menu_forms() should not duplicate menu creation for DMP and non-DMP apps
and should rely on the Application to have already done that.  It can simply
use the cached menu structure.  This will also work with minor changes in the
'new-commands' branch.
2) The only other use of get_menu_forms() is actually in application.py
checking for existence of startup forms in menu.  Checking form classes there seems
redundant and will not work in the 'new-commands' branch anyway so we cen use
'get_menu_defs() instead of 'get_menu_forms()' and get rid of
'get_menu_forms()' altogether.

In addition 'app.run()' calls are updated to use 'over' to simplify the code
and avoid Python 2/3 compatibility handling of division.