Add support for buttons with both text and icon
The previous implementation only supported buttons with either text label or an
icon.  Using an icon lead to dismissing the label.  So the label, if specified,
was only a fallback for the case that the icon is not found.  As we rarely want
such unpredictable results, this logic has been changed and all existing
wx_button calls were checked to comply with the new logic.  Only one of the
arguments is used in order to achieve the same results as before.