Allow several directories to be searched for icons
This is practical to allow applications to include their own icons and use them
for their user interface, for example forms and dialogs.  We don't want the
applications to rely on Pytis icons as we have no control on which of them are
used and cannot tie our hands changing the available icon set.