Avoid ExposedFileWrapper encoding/mode mismatch in Python 3
If mode is 'w' and encoding is specified, exposed_write() raises "TypeError:
write() argument must be str, not bytes" inc Python 3.  This is because if we
manually encode the string before writing, it (not surprisingly) becomes bytes.
Python 2 is tolerant to this inconsistency so we rather don't touch Python 2
behavior.