Don't allow application defined custom commands
Instead of defining a custom command, the application can achieve the same by
using an existing command 'app.call()'.  The application defined commands
didn't actually work with keyboard shortcuts because
'pytis.presentation.Application' is not a KeyHandler instance so its commands
in the keymap were ignored as only KeyHandler defined commands are processed
and is is not possible to use the 'pytis.presentation.Application' as a key
guardian.