Replace numerous text editor solutions with a single one
We had three classes to implement a text editor with a resizable text field.
There was ResizableEditForm, ResizableInputForm and StructuredTextEditor.  They
used a specific form construction
implementation (ResizableEditForm._create_form_parts()) with very limited
capabilities in order to make the text field resizable.  This method didn't
support standard layout options, just created a single unlabeled field.  Its
code was a simplified duplication of other method's code in the parent class.
It was not possible to use such forms generally and not through Pytis API.

The new approach makes it possible to create resizable text fields through
app.input_text().  The field is automatically made resizable, when it is
multiline (height > 1).  It is possible to create such fields with a
label (compact or standard) or unlabeled (which often looks more natural).  In
order to replace previous StructuredTextEditor, app.input_text() now supports
the new arguments text_format and attachment_storage with the same meaning as
in Field specification.

The new implementation allows constuction of resizable fields within the base
class by adding a new keyword argument 'growable' to '_create_group_panel()',
'_create_group()' and '_pack_fields()'.  When 'growable' is False (which it
always is in the base class), the form construction is exactly the same as
before.  This avoids potential disruption of existing form layouts.  In many
places, the layout would probably not change even if we used the same code for
both cases of the 'growable' value, but it would have to be tested to verify
that assumption.  For now I decided for a very conservative aproach.