Fix switching tabs invoked from web content links
When a form is opened through a link inside a WebForm, it is handled as an
event of the built in browser (our class Browser in screen.py).  When
run_form() is called from an event handler, the events emited within it are
suppressed (as nesting events is not supported in event.py).  Calling
run_form() through wxCallAfter resolves the situation and run_form events are
handled.

This fixes the problem when a form was opened from a welcome page link and this
form was already present in existing tabs.  The tab was switched in this case,
but the form was not activated (stayed gray), because
wx.aui.EVT_AUINOTEBOOK_PAGE_CHANGED was suppressed and thus
Application._on_page_change() did not get called.