Avoid print menu construction in _can_print_menu()
Menu construction causes database errors in pytis.form unit tests because the
related database objects (printing.UserOutputTemplates) don't exist in the test
environment.  To avoid bloating test environment, it seems more reasonable to
just decide upon print rights as this actually is the primary condition which
causes the print menu to exist or not.