Fix column resizing in recent wxPython versions
Column resizing worked in older wxPython versions used with Python 2.7 (such
wxPython 4.1.0), but the newer versions used with Python 3 (such as wxPython
4.2.1), it stopped to work, showing a traceback like the following in the
background:

  wx._core.wxAssertionError: C++ assertion ""m_dragRowOrCol != -1""
  failed at /home/wxpy/wxPython-4.2.1/ext/wxWidgets/src/generic/grid.cpp(5057)
  in DoGridDragResize(): shouldn't be called when not drag resizing

  The above exception was the direct cause of the following exception:

  SystemError: <class 'wx._html2.WebViewEvent'> returned a result with an error set

A bit of trial revealed that the problem is caused by our wx.EVT_LEFT_DOWN
mouse event handler which doesn't call `event.Skip()`.  Adding this call fixes
the problem.

The previous comment indicates that the call to Skip() was left out
deliberately in order to prevent range selection.  As this event handler is
limited to column label area, it is unclear what is exactly meant by "range
selection".  Even with the Skip() call in place it is not possible to perform
any sort of unexpected selection by clicking or dragging column labels (with or
without Shift pressed), so this comment was hopefully outdated and referred to
some previous setup or behavior which is not relevant any more.