Fix remote writes with pd.Binary.Data instances
When passing pytis.data.Binary.Data instances directly to write_file() or
remote file's write method, the remote write fails with

TypeError: a bytes-like object is required, not 'pytis.data.types_.Data'

This is because the related netref instance passed to the remote side is not
recognized as a bytes instance any more.

I tried to make RPyC to convert pytis.data.Binary.Data instances to bytes
automatically, but without success.  Thus this hack seems still the cleanest
way to fix the problem.