Allow creating data object right from db specs
In applications it is often necessary to work with database tables.  Up to now,
we were always using the form specification name to create a data object using
pytis.util.data_object('spec.Name').  But this very often leads to the need to
define form specifications solely for the purpose of creating data objects.
These specifications are typically named '*Proc' and have the attribute
'public' set to false in order to avoid creating forms from them.  This does
not make much sense, because form specifications exist in order to define
forms, not tables.

This commit makes it possible to use gensqlalchemy table specifications
directly as an argument to 'pytis.util.data_object()'.  This avoids the need to
define extra non-public form specifications just for that purpose.  This
simplifies maintenance and removes one level of indirection which makes the
code hardly readable.